// packages/wallet/src/lib/types/nockblocks-api-v1.ts
export interface components {
  schemas: {
// ...
    BlockObjectV1: {
      /** @description Block version number */
      version: number;
      /** @description Block hash/digest as base58 string */
      digest: string;
      /**
       * Format: uri
       * @description URL to the proof for this block
       */
      proofUrl: string;
      /** @description Parent block hash as base58 string */
      parent: string;
      /** @description Array of transaction IDs (base58) */
      txids: string[];
      /** @description Array of coinbase note objects */
      coinbaseNotes: components['schemas']['NoteObjectV1'][];
      /**
       * Format: int64
       * @description Block timestamp in Unix seconds
       */
      timestamp: number;
      /**
       * Format: int64
       * @description Counter within the current epoch
       */
      epochCounter: number;
      /** @description Total accumulated work as string */
      accumulatedWork: string;
      /** @description Mining target as string */
      target: string;
      /**
       * Format: int64
       * @description Block height in the chain
       */
      height: number;
      /** @description Optional message included in the block */
      msg: string;
      /** @description Optional array of full transaction objects (only if includeTransactions was true) */
      transactionsData?: components['schemas']['TransactionObjectV1'][] | null;
    };
    NoteObjectV1: components['schemas']['NoteV0Object'] | components['schemas']['NoteV1Object'];
// ...
    TransactionObjectV1:
      | components['schemas']['TransactionV0Object']
      | components['schemas']['TransactionV1Object'];
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}